requirements.txt
torch>=2.1.0
transformers>=4.40.0
datasets>=2.18.0
sentence-transformers>=2.7.0
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0
pandas>=2.2.0
pyyaml>=6.0
tqdm>=4.66.0
accelerate>=0.28.0
evaluate>=0.4.1
rouge-score>=0.1.2
matplotlib>=3.8.0
seaborn>=0.13.0
sympy>=1.12
jsonlines>=4.0.0
typer>=0.12.0
rich>=13.7.0
